+++ PicosAdventure/Utils/textClass.h
@@ -28,13 +28,13 @@ public:
 	TextClass(const TextClass&);
 	~TextClass();
 
-	bool setup(ID3D11Device* device, ID3D11DeviceContext* deviceContext, Shader2DClass* shader, int screenWidth, int screenHeight, int posX, int posY, std::string sentenceText);
+	bool setup(ID3D11Device* device, ID3D11DeviceContext* deviceContext, Shader2DClass* shader, int screenWidth, int screenHeight, float posX, float posY, std::string sentenceText);
 	bool draw(ID3D11DeviceContext* deviceContext, XMFLOAT4X4 worldMatrix, XMFLOAT4X4 baseViewMatrix, XMFLOAT4X4 orthoMatrix);
 	void destroy();
 
 	bool setText(std::string text, ID3D11DeviceContext* deviceContext);
 
-	void setPosition(int positionX, int positionY);
+	void setPosition(float positionX, float positionY);
 
 	Point getTextPosition();
 	int getTextWidth();
@@ -42,7 +42,7 @@ public:
 
 private:
 	bool setupSentence(SentenceType** sentence, int maxLength, ID3D11Device* device);
-	bool updateSentence(SentenceType* sentence, std::string text, int positionX, int positionY, float red, float green, float blue, ID3D11DeviceContext* deviceContext);
+	bool updateSentence(SentenceType* sentence, std::string text, float positionX, float positionY, float red, float green, float blue, ID3D11DeviceContext* deviceContext);
 	bool drawSentence(ID3D11DeviceContext* deviceContext, SentenceType* sentence, XMFLOAT4X4 worldMatrix, XMFLOAT4X4 baseViewMatrix, XMFLOAT4X4 orthoMatrix);
 	void destroySentence(SentenceType** sentence);
 
@@ -50,7 +50,7 @@ private:
 	FontClass* font_;
 	Shader2DClass* fontShader_;
 	int screenWidth_, screenHeight_;
-	int positionX_, positionY_;
+	float positionX_, positionY_;
 	SentenceType* sentence_;
 	std::string text_;
 };
